This program is "Four in a Row"
The was originally typed into a Timex Sinclair  2068, and debugged. Then translated into BBC Basic, which is the current form of the file.
It seems to work fine, maybe too fine, cause I never win...  
This will work perfectly in the AgonLight2, and that is why I am putting it in a public repository.